+++ lib/peddler/headers.rb
@@ -47,7 +47,10 @@ module Peddler
       match_data = headers['Content-Type']&.match(/charset=(.*);?/)
       return unless match_data
 
-      Encoding.find(match_data[1])
+      string = match_data[1]
+      string = 'UTF-8' if string == 'UTF8'
+
+      Encoding.find(string)
     end
 
     # The max hourly request quota for the requested operation